rpios-testing/rfid/requirements.txt
mfrc522-python
